+++ bgameb/__init__.py
@@ -1,6 +1,7 @@
 from bgameb.games import Game
 from bgameb.players import Human, Bot
 from bgameb.rollers import Dice, Coin
+from bgameb.cards import Card
 from bgameb.shakers import Shaker
 from bgameb.utils import log_enable
 from bgameb._version import __version__