+++ bgameb/cards.py
@@ -1,0 +1,100 @@
+"""Game cards
+"""
+from typing import Optional, Any
+from dataclasses import dataclass, field
+from dataclasses_json import DataClassJsonMixin
+from bgameb.utils import log_me
+
+
+class CardTexts(dict):
+    """Cards texts collection
+    """
+    def __init__(self, /, **kwargs) -> None:
+        self.__dict__.update(kwargs)
+
+    def __getattr__(self, attr: str) -> str:
+        try:
+            return self[attr]
+        except KeyError:
+            raise AttributeError(attr)
+
+    def __setattr__(self, attr: str, value: str) -> None:
+        self[attr] = value
+
+    def __delattr__(self, attr: str) -> None:
+        del self[attr]
+
+    def __repr__(self):
+        items = (f"{k}={v!r}" for k, v in self.items())
+        return "{}({})".format(type(self).__name__, ", ".join(items))
+
+    def __eq__(self, other: Any) -> bool:
+        if isinstance(self, dict) and isinstance(other, dict):
+            return self.__dict__ == other.__dict__
+        return NotImplemented
+
+
+@dataclass
+class Card(DataClassJsonMixin):
+    """Create the card
+    """
+    name: str = 'card'
+    open: bool = False
+    tapped: bool = False
+    side: Optional[str] = None
+    text: CardTexts = field(default_factory=dict, init=False)
+
+    def __post_init__(self) -> None:
+        self.text = CardTexts()
+
+        # set logger
+        self.logger = log_me.bind(
+            classname=self.__class__.__name__,
+            name=self.name)
+        self.logger.info(f'Card created.')
+
+    def flip(self) -> None:
+        """Face up or face down the card regardles of it condition
+        """
+        if self.open:
+            self.open = False
+            self.logger.debug(f'Card face down.')
+        else:
+            self.open = True
+            self.logger.debug(f'Card face up.')
+
+    def face_up(self) -> CardTexts:
+        """Face up the card and return text
+        """
+        self.open = True
+        self.logger.debug(f'Card face up.')
+        return self.text
+
+    def face_down(self) -> None:
+        """Face down the card
+        """
+        self.open = False
+        self.logger.debug(f'Card face down.')
+
+    def tap(self, side='right') -> None:
+        """Tap the card to the given side
+        """
+        self.tapped = True
+        self.side = side
+        self.logger.debug(f'Card taped to side {side}.')
+
+    def untap(self) -> None:
+        """Untap the card
+        """
+        self.tapped = False
+        self.logger.debug(f'Card untaped.')
+
+    def alter(self) -> None:
+        """Many cards have alter views. For example
+        card can have main view, that apply most time of the game
+        and second view, that apply only if card played as
+        that alternative. For ease of understanding, consider that
+        different views of the same card are not related directly
+        to each other.
+        """
+        raise NotImplementedError